# express/static/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"securitySchemes": {
			"keycloak": {
				"type": "apiKey",
				"name": "Authorization",
				"description": "Keycloak Bearer Token",
				"in": "header"
			}
		}
	},
	"info": {
		"title": "backend",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
		"/products": {
			"get": {
				"operationId": "GetProducts",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"properties": {
											"UnitsOnOrder": {
												"type": "number",
												"format": "double"
											},
											"UnitsInStock": {
												"type": "number",
												"format": "double"
											},
											"UnitPrice": {
												"type": "string"
											},
											"SupplierId": {
												"type": "number",
												"format": "double"
											},
											"ReorderLevel": {
												"type": "number",
												"format": "double"
											},
											"QuantityPerUnit": {
												"type": "string"
											},
											"Discontinued": {
												"type": "number",
												"format": "double"
											},
											"CategoryId": {
												"type": "string"
											},
											"ProductName": {
												"type": "string"
											},
											"Id": {
												"type": "number",
												"format": "double"
											}
										},
										"required": [
											"UnitsOnOrder",
											"UnitsInStock",
											"UnitPrice",
											"SupplierId",
											"ReorderLevel",
											"QuantityPerUnit",
											"Discontinued",
											"CategoryId",
											"ProductName",
											"Id"
										],
										"type": "object"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "CreateProduct",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"UnitsOnOrder": {
											"type": "number",
											"format": "double"
										},
										"UnitsInStock": {
											"type": "number",
											"format": "double"
										},
										"UnitPrice": {
											"type": "string"
										},
										"SupplierId": {
											"type": "number",
											"format": "double"
										},
										"ReorderLevel": {
											"type": "number",
											"format": "double"
										},
										"QuantityPerUnit": {
											"type": "string"
										},
										"Discontinued": {
											"type": "number",
											"format": "double"
										},
										"CategoryId": {
											"type": "string"
										},
										"ProductName": {
											"type": "string"
										},
										"Id": {
											"type": "number",
											"format": "double"
										}
									},
									"required": [
										"UnitsOnOrder",
										"UnitsInStock",
										"UnitPrice",
										"SupplierId",
										"ReorderLevel",
										"QuantityPerUnit",
										"Discontinued",
										"CategoryId",
										"ProductName",
										"Id"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"security": [
					{
						"keycloak": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"UnitsOnOrder": {
										"type": "number",
										"format": "double"
									},
									"UnitsInStock": {
										"type": "number",
										"format": "double"
									},
									"UnitPrice": {
										"type": "string"
									},
									"SupplierId": {
										"type": "number",
										"format": "double"
									},
									"ReorderLevel": {
										"type": "number",
										"format": "double"
									},
									"QuantityPerUnit": {
										"type": "string"
									},
									"Discontinued": {
										"type": "number",
										"format": "double"
									},
									"CategoryId": {
										"type": "string"
									},
									"ProductName": {
										"type": "string"
									}
								},
								"required": [
									"UnitsOnOrder",
									"UnitsInStock",
									"UnitPrice",
									"SupplierId",
									"ReorderLevel",
									"Discontinued",
									"CategoryId",
									"ProductName"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/products/{id}": {
			"get": {
				"operationId": "GetProductById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"properties": {
												"UnitsOnOrder": {
													"type": "number",
													"format": "double"
												},
												"UnitsInStock": {
													"type": "number",
													"format": "double"
												},
												"UnitPrice": {
													"type": "string"
												},
												"SupplierId": {
													"type": "number",
													"format": "double"
												},
												"ReorderLevel": {
													"type": "number",
													"format": "double"
												},
												"QuantityPerUnit": {
													"type": "string"
												},
												"Discontinued": {
													"type": "number",
													"format": "double"
												},
												"CategoryId": {
													"type": "string"
												},
												"ProductName": {
													"type": "string"
												},
												"Id": {
													"type": "number",
													"format": "double"
												}
											},
											"required": [
												"UnitsOnOrder",
												"UnitsInStock",
												"UnitPrice",
												"SupplierId",
												"ReorderLevel",
												"QuantityPerUnit",
												"Discontinued",
												"CategoryId",
												"ProductName",
												"Id"
											],
											"type": "object"
										},
										{
											"type": "string",
											"enum": [
												"ไม่พบข้อมูล"
											]
										}
									]
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"patch": {
				"operationId": "UpdateProduct",
				"responses": {
					"200": {
						"description": "Updated",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"UnitsOnOrder": {
											"type": "number",
											"format": "double"
										},
										"UnitsInStock": {
											"type": "number",
											"format": "double"
										},
										"UnitPrice": {
											"type": "string"
										},
										"SupplierId": {
											"type": "number",
											"format": "double"
										},
										"ReorderLevel": {
											"type": "number",
											"format": "double"
										},
										"QuantityPerUnit": {
											"type": "string"
										},
										"Discontinued": {
											"type": "number",
											"format": "double"
										},
										"CategoryId": {
											"type": "string"
										},
										"ProductName": {
											"type": "string"
										},
										"Id": {
											"type": "number",
											"format": "double"
										}
									},
									"required": [
										"UnitsOnOrder",
										"UnitsInStock",
										"UnitPrice",
										"SupplierId",
										"ReorderLevel",
										"QuantityPerUnit",
										"Discontinued",
										"CategoryId",
										"ProductName",
										"Id"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"security": [
					{
						"keycloak": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"UnitsOnOrder": {
										"type": "number",
										"format": "double"
									},
									"UnitsInStock": {
										"type": "number",
										"format": "double"
									},
									"UnitPrice": {
										"type": "string"
									},
									"SupplierId": {
										"type": "number",
										"format": "double"
									},
									"ReorderLevel": {
										"type": "number",
										"format": "double"
									},
									"QuantityPerUnit": {
										"type": "string"
									},
									"Discontinued": {
										"type": "number",
										"format": "double"
									},
									"CategoryId": {
										"type": "string"
									},
									"ProductName": {
										"type": "string"
									}
								},
								"type": "object"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteProduct",
				"responses": {
					"200": {
						"description": "Deleted",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"security": [
					{
						"keycloak": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}